+++ terraform/variables.tf
@@ -2,10 +2,6 @@ variable "mongodb" {
   type = map(string)
 }
 
-variable "formio" {
-  type = map(string)
-}
-
 variable "host" {
   type = string
 }